+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Avoid creating ShipEffects VesselModule for debris, spaceobjects, and dropped items - thanks @bimo1d
 - Improve performance when vessel part count changes several times in a frame (e.g. during RUD)
 - Avoid creating ShipEffects VesselModule for non-loaded vessels
+- Cache references to mixer groups
 
 ## 0.9.13 - 2026-04-20